Discrete block library

Index

ModelingToolkitSampledData.ClockChanger

ModelingToolkitSampledData.Delay

ModelingToolkitSampledData.Difference

ModelingToolkitSampledData.DiscreteDerivative

ModelingToolkitSampledData.DiscreteIntegrator

ModelingToolkitSampledData.DiscretePIDParallel

ModelingToolkitSampledData.DiscretePIDStandard

ModelingToolkitSampledData.NormalNoise

ModelingToolkitSampledData.Sampler

ModelingToolkitSampledData.UniformNoise

ModelingToolkitSampledData.ZeroOrderHold

Docstrings

ModelingToolkitSampledData.ClockChanger — Constant

ClockChanger()
ClockChanger(; to::AbstractClock, from::AbstractClock)

Connectors:

input (continuous-time signal)

output (discrete-time signal)

Experimental

The ClockChanger component is experimental and has known correctness issues. Please use with caution.

ModelingToolkitSampledData.Delay — Constant

Delay(; n = 1)

A discrete-time delay of n samples, corresponding to the transfer function $z^{-n}$.

Connectors:

input

output

Structural Parameters:

n: Number of delay samples

ModelingToolkitSampledData.Difference — Constant

Difference()

A discrete-time difference operator, corresponding to the transfer function $1 - z^{-1}$.

Connectors:

input

output

For a discrete-time finite-difference derivative approximation, see DiscreteDerivative.

ModelingToolkitSampledData.DiscreteDerivative — Constant

DiscreteDerivative(; k = 1)

A discrete-time derivative filter, corresponding to the transfer function $k (z-1) / (T_s z)$

where $T_s$ is the sample time of the derivative filter.

Connectors:

input

output

Parameters:

k: Gain of derivative filter

ModelingToolkitSampledData.DiscreteIntegrator — Constant

DiscreteIntegrator(;name, k = 1, x = 0.0, method = :backward)

Outputs y = ∫k*u dt, discretized to correspond to the one of the discrete-time transfer functions

method = :forward: $T_s / (z - 1)$

method = :backward: $T_s z / (z - 1)$

method = :trapezoidal: $(T_s / 2) (z + 1) / (z - 1)$

where $T_s$ is the sample time of the integrator.

Initial value of integrator state $x$ can be set with x

Connectors:

input

output

Parameters:

k: Gain of integrator

ModelingToolkitSampledData.DiscretePIDParallel — Constant

DiscretePIDParallel(;name, kp = 1, ki = 1, kd = 1, Ni = √(max(kd * ki, 1e-6)), Nd = 10kp, u_max = Inf, u_min = -u_max, wp = 1, wd = 1, Ts = 1, with_I = true, with_D = true, Imethod = :forward, Dmethod = :backward)

Discrete-time PID controller on parallel form with anti-windup and set-point weighting.

The controller is implemented on parallel form:

Simplified:

\[u = k_p e + \int k_i e dt + k_d \dfrac{de}{dt} \]

Detailed:

\[u = k_p(w_p r - y) + \int \big( k_i (r - y) + N_i e_s \big ) dt + k_d \dfrac{d}{dt}(w_d r - y)\]

where e_s = u - v is the saturated error signal, v is the unsaturated control signal and u is the saturated control signal.

The derivative is filtered to allow a maximum gain of $N_d$.

The integrator is discretized using the method specified by Imethod, options include

Imethod = :forward (default): Corresponding to the transfer function $T_s / (z - 1)$

Imethod = :backward: Corresponding to the transfer function $T_s z / (z - 1)$

Imethod = :trapezoidal: Corresponding to the transfer function $(T_s / 2) (z + 1) / (z - 1)$

The derivative is discretized using the method specified by Dmethod, options include

Dmethod = :forward: Corresponding to the transfer function $\dfrac{N (z-1)}{z - \dfrac{k_d-N T_s}{k_d}}$.

Dmethod = :backward (default): Corresponding to the transfer function $\dfrac{\dfrac{Nk_d}{k_d + N T_s}(z-1)}{z - \dfrac{k_d}{k_d + N T_s}}$

Anti windup is realized by tracking using the gain $N_i$ on the error signal $e_s$ when the output is saturated.

To use the controller in 1DOF mode, i.e., with only the control error as input, connect the error signal to the reference connector, connect a Constant(; k = 0) to the measurement connector and set wp = wd = 1.

Connectors:

reference: The reference signal to the controller (or the error signal if used in 1DOF mode)

measurement: The measurement feedback

ctr_output: The control signal output

Parameters:

kp: Proportional gain

ki: Integral gain (only active if with_I = true)

kd: Derivative gain (only active if with_D = true)

Ni: Anti-windup gain (only active if with_I = true)

Nd: Maximum derivative gain (only active if with_D = true). Typically set to 10-100 times the proportional gain.

u_max: Maximum output above which the output is saturated

u_min: Minimum output below which the output is saturated. This defaults to -u_max if u_max > 0 and -Inf otherwise.

wp: [0, 1] Set-point weighting in the proportional part. Set to 0 to prevent step changes in the output due to step changes in the reference.

wd: [0, 1] Set-point weighting in the derivative part. Set to 0 to prevent very large impulsive changes in the output due to step changes in the reference.

with_I: Whether or not to include the integral part

with_D: Whether or not to include the derivative part

Imethod: Discretization method for the integrator (see details above)

Dmethod: Discretization method for the derivative (see details above)

Extended help:

Internal variables:

I: State of integrator

D: State of filtered derivative

r: Reference signal internal variable

y: Measurement signal internal variable

wde: Setpoint-weighted error for derivative

v: Un-saturated output of the controller

u: Saturated output of the controller

eI: Error signal input to integrator including anit-windup tracking signal

e: Error signal

ModelingToolkitSampledData.DiscretePIDStandard — Constant

DiscretePIDStandard(;name, K = 1, Ti = 1, Td = 1, Ni = √(max(kd * ki, 1e-6)), Nd = 10, u_max = Inf, u_min = -u_max, wp = 1, wd = 1, Ts = 1, with_I = true, with_D = true, Imethod = :forward, Dmethod = :backward)

Discrete-time PID controller on standard (ideal) form with anti-windup and set-point weighting.

The controller is implemented on standard form

Simplified:

\[u = K \left( e + \int \frac{1}{T_i} e dt + T_d \dfrac{de}{dt} \right)\]

Detailed:

\[u = K \left( (w_p r - y) + \int \big( \frac{1}{T_i} (r - y) + N_i e_s \big ) dt + T_d \dfrac{d}{dt}(w_d r - y) \right)\]

where e_s = u - v is the saturated error signal, v is the unsaturated control signal and u is the saturated control signal.

The derivative is filtered to allow a maximum gain of $N_d$.

The integrator is discretized using the method specified by Imethod, options include

Imethod = :forward (default): Corresponding to the transfer function $T_s / (z - 1)$

Imethod = :backward: Corresponding to the transfer function $T_s z / (z - 1)$

Imethod = :trapezoidal: Corresponding to the transfer function $(T_s / 2) (z + 1) / (z - 1)$

The derivative is discretized using the method specified by Dmethod, options include

Dmethod = :forward: Corresponding to the transfer function $\dfrac{N (z-1)}{z - \dfrac{T_d-N T_s}{T_d}}$.

Dmethod = :backward (default): Corresponding to the transfer function $\dfrac{\dfrac{NT_d}{T_d + N T_s}(z-1)}{z - \dfrac{T_d}{T_d + N T_s}}$

Anti windup is realized by tracking using the gain $N_i$ on the error signal $e_s$ when the output is saturated.

To use the controller in 1DOF mode, i.e., with only the control error as input, connect the error signal to the reference connector, connect a Constant(; k = 0) to the measurement connector and set wp = wd = 1.

Connectors:

reference: The reference signal to the controller (or the error signal if used in 1DOF mode)

measurement: The measurement feedback

ctr_output: The control signal output

Parameters:

K: Proportional gain

Ti: Integral time constant (only active if with_I = true)

Td: Derivative time (only active if with_D = true)

Ni: Anti-windup gain (only active if with_I = true)

Nd: Maximum derivative gain (only active if with_D = true). Typically set to 10-100.

u_max: Maximum output above which the output is saturated

u_min: Minimum output below which the output is saturated. This defaults to -u_max if u_max > 0 and -Inf otherwise.

wp: [0, 1] Set-point weighting in the proportional part. Set to 0 to prevent step changes in the output due to step changes in the reference.

wd: [0, 1] Set-point weighting in the derivative part. Set to 0 to prevent very large impulsive changes in the output due to step changes in the reference.

with_I: Whether or not to include the integral part

with_D: Whether or not to include the derivative part

Imethod: Discretization method for the integrator (see details above)

Dmethod: Discretization method for the derivative (see details above)

Extended help:

Internal variables:

I: State of integrator

D: State of filtered derivative

r: Reference signal internal variable

y: Measurement signal internal variable

wde: Setpoint-weighted error for derivative

v: Un-saturated output of the controller

u: Saturated output of the controller

eI: Error signal input to integrator including anit-windup tracking signal

e: Error signal

ModelingToolkitSampledData.NormalNoise — Constant

NormalNoise()

A discrete-time noise source that returns a normally-distributed value at each clock tick.

Parameters

mu = 0: Mean

sigma = 1: Standard deviation

Structural parameters

rng: A random number generator, defaults to Random.Xoshiro().

z: The ShiftIndex used to indicate clock partition.

Connectors:

output

ModelingToolkitSampledData.Sampler — Constant

Sampler()
Sampler(; dt::Real)
Sampler(; clock::AbstractClock)

Sampler transforms a continuous-time signal into discrete time by sampling the input signal every time the associated clock ticks. The clock can be specified explicitly using the clock keyword argument, or implicitly by providing a sample time dt, in which case a standard periodic Clock is used.

Connectors:

input (continuous-time signal)

output (discrete-time signal)

ModelingToolkitSampledData.UniformNoise — Constant

UniformNoise()

A discrete-time noise source that returns a uniformly distributed value at each clock tick.

Parameters

l = 0: Lower bound

u = 1: Upper bound

Structural parameters

rng: A random number generator, defaults to Random.Xoshiro().

z: The ShiftIndex used to indicate clock partition.

Connectors:

output

ModelingToolkitSampledData.ZeroOrderHold — Constant

ZeroOrderHold()

Zero-order-Hold translates a discrete time (clocked) signal into continuous time by holding the value of the discrete signal constant until the next sample.

Connectors:

input (discrete-time signal)

output (continuous-time signal)